requirement.txt
python 3.6.10
pytorch 1.5.0
torchvision 0.6.0
cudatoolkit 10.2.89
cudnn 7.6.5
